Out of time for the day, refactored the index and customer page.
It seems to only like the .jsp file extensions but they work exactly the
same as html anyways, i'll move the tables into directives once we know
about what data we will be using later. The actual mvc needs to be
implemented in a bit for actual redirects which should be very possible
with the current project structure, can probably be done between
customer and index right now, pretty sure the index works unless i'm
missing a couple of imports still.